Refactor DType import and usage across modules
- Move DType enum to new types.py module
- Update import statements in __init__.py, distribution.py, and random.py
- Standardize DType enum values (Float, Int) for consistency

+++ py-diffusionx/python/diffusionx/types.py
@@ -1,6 +1,5 @@
 from enum import Enum
 
-
 class DType(Enum):
-    INT = "int"
-    FLOAT = "float"
+    Float = "float"
+    Int = "int"
